+++ packages/onchainkit/src/checkout/components/Checkout.tsx
@@ -1,11 +1,15 @@
 'use client';
 
 import { useIsMounted } from '../../internal/hooks/useIsMounted';
-import { useTheme } from '../../internal/hooks/useTheme';
 import { cn } from '../../styles/theme';
-import type { CheckoutReact } from '../types';
+import type { CheckoutProps } from '../types';
 import { CheckoutProvider } from './CheckoutProvider';
 
+/**
+ * @deprecated The <Checkout /> component and its related components and hooks are deprecated
+ * and will be removed in a future version. We recommend looking at Base Pay for similar functionality.
+ * @see {@link https://docs.base.org/base-account/guides/accept-payments}
+ */
 export function Checkout({
   chargeHandler,
   children,
@@ -13,9 +17,8 @@ export function Checkout({
   isSponsored,
   onStatus,
   productId,
-}: CheckoutReact) {
+}: CheckoutProps) {
   const isMounted = useIsMounted();
-  const componentTheme = useTheme();
   // prevents SSR hydration issue
   if (!isMounted) {
     return null;
@@ -28,9 +31,7 @@ export function Checkout({
       onStatus={onStatus}
       productId={productId}
     >
-      <div
-        className={cn(componentTheme, 'flex w-full flex-col gap-2', className)}
-      >
+      <div className={cn('flex w-full flex-col gap-2', className)}>
         {children}
       </div>
     </CheckoutProvider>