+++ packages/onchainkit/src/fund/components/FundCard.tsx
@@ -1,10 +1,9 @@
 'use client';
 
 import type { ReactNode } from 'react';
-import { useTheme } from '../../internal/hooks/useTheme';
-import { background, border, cn, color, text } from '../../styles/theme';
+import { border, cn, text } from '../../styles/theme';
 import { useFundCardSetupOnrampEventListeners } from '../hooks/useFundCardSetupOnrampEventListeners';
-import type { FundCardPropsReact } from '../types';
+import type { FundCardProps } from '../types';
 import FundCardAmountInput from './FundCardAmountInput';
 import FundCardAmountInputTypeSwitch from './FundCardAmountInputTypeSwitch';
 import { FundCardHeader } from './FundCardHeader';
@@ -15,6 +14,7 @@ import { FundCardSubmitButton } from './FundCardSubmitButton';
 
 export function FundCard({
   assetSymbol,
+  sessionToken,
   buttonText = 'Buy',
   headerText,
   country = 'US',
@@ -26,9 +26,11 @@ export function FundCard({
   onError,
   onStatus,
   onSuccess,
-}: FundCardPropsReact) {
-  const componentTheme = useTheme();
-
+}: FundCardProps) {
+  // Ensure a session token is provided
+  if (!sessionToken) {
+    throw new Error('FundCard requires a sessionToken');
+  }
   return (
     <FundCardProvider
       asset={assetSymbol}
@@ -41,15 +43,14 @@ export function FundCard({
       onStatus={onStatus}
       onSuccess={onSuccess}
       presetAmountInputs={presetAmountInputs}
+      sessionToken={sessionToken}
     >
       <div
         className={cn(
-          componentTheme,
-          background.default,
-          color.foreground,
-          'flex w-full flex-col p-6',
+          'bg-ock-background',
+          'text-ock-foreground',
+          'rounded-ock-default flex w-full flex-col p-6',
           text.headline,
-          border.radius,
           border.lineDefault,
           className,
         )}