+++ packages/onchainkit/src/minikit/MiniKitProvider.tsx
@@ -1,10 +1,5 @@
 'use client';
-
-import { DefaultOnchainKitProviders } from '@/DefaultOnchainKitProviders';
-import { OnchainKitProvider } from '@/OnchainKitProvider';
-import type { OnchainKitProviderReact } from '@/types';
-import sdk, { type Context } from '@farcaster/frame-sdk';
-import { farcasterFrame } from '@farcaster/frame-wagmi-connector';
+import sdk, { type Context } from '@farcaster/miniapp-sdk';
 import {
   createContext,
   useCallback,
@@ -12,45 +7,65 @@ import {
   useMemo,
   useState,
 } from 'react';
-import { coinbaseWallet } from 'wagmi/connectors';
 import type {
   MiniKitContextType,
-  MiniKitProviderReact,
+  MiniKitProviderProps,
   UpdateClientContextParams,
 } from './types';
+import { AutoConnect } from './components/AutoConnect';
 
-export const emptyContext = {} as MiniKitContextType;
+export const MiniKitContext = createContext<MiniKitContextType>({
+  enabled: false,
+  context: null,
+  updateClientContext: () => {},
+  notificationProxyUrl: '',
+  __isMiniKit: false,
+});
 
-export const MiniKitContext = createContext<MiniKitContextType>(emptyContext);
-
-/**
- * Provides the MiniKit React Context to the app.
- */
-export function MiniKitProvider({
+function MiniKitProviderContent({
   children,
   notificationProxyUrl = '/api/notify',
-  ...onchainKitProps
-}: MiniKitProviderReact & OnchainKitProviderReact) {
-  const [context, setContext] = useState<Context.FrameContext | null>(null);
+  autoConnect = true,
+}: MiniKitProviderProps) {
+  const [context, setContext] = useState<Context.MiniAppContext | null>(null);
+
+  const updateClientContext = useCallback(
+    ({ details, miniAppAdded }: UpdateClientContextParams) => {
+      setContext((prevContext) => {
+        if (!prevContext) {
+          return null;
+        }
+        return {
+          ...prevContext,
+          client: {
+            ...prevContext.client,
+            notificationDetails: details ?? undefined,
+            added: miniAppAdded ?? prevContext.client.added,
+          },
+        };
+      });
+    },
+    [],
+  );
 
   useEffect(() => {
-    sdk.on('frameAdded', ({ notificationDetails }) => {
+    sdk.on('miniAppAdded', ({ notificationDetails }) => {
       if (notificationDetails) {
         updateClientContext({
           details: notificationDetails,
-          frameAdded: true,
+          miniAppAdded: true,
         });
       }
     });
 
-    sdk.on('frameAddRejected', ({ reason }) => {
-      console.error('Frame add rejected', reason);
+    sdk.on('miniAppAddRejected', ({ reason }) => {
+      console.error('MiniApp add rejected', reason);
     });
 
-    sdk.on('frameRemoved', () => {
+    sdk.on('miniAppRemoved', () => {
       updateClientContext({
         details: undefined,
-        frameAdded: false,
+        miniAppAdded: false,
       });
     });
 
@@ -81,64 +96,41 @@ export function MiniKitProvider({
     return () => {
       sdk.removeAllListeners();
     };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  const updateClientContext = useCallback(
-    ({ details, frameAdded }: UpdateClientContextParams) => {
-      setContext((prevContext) => {
-        if (!prevContext) {
-          return null;
-        }
-        return {
-          ...prevContext,
-          client: {
-            ...prevContext.client,
-            notificationDetails: details ?? undefined,
-            added: frameAdded ?? prevContext.client.added,
-          },
-        };
-      });
-    },
-    [],
-  );
-
-  const connectors = useMemo(() => {
-    return [
-      context // if context is set, the app is running in a frame, use farcasterFrame connector
-        ? farcasterFrame()
-        : coinbaseWallet({
-            appName: process.env.NEXT_PUBLIC_ONCHAINKIT_PROJECT_NAME,
-            appLogoUrl: process.env.NEXT_PUBLIC_ICON_URL,
-            preference: 'all',
-          }),
-    ];
-  }, [context]);
+  }, [updateClientContext]);
 
   const value = useMemo(() => {
     return {
+      enabled: true,
       context,
       updateClientContext,
       notificationProxyUrl,
+      __isMiniKit: true,
     };
   }, [updateClientContext, notificationProxyUrl, context]);
 
   return (
     <MiniKitContext.Provider value={value}>
-      <DefaultOnchainKitProviders connectors={connectors}>
-        <OnchainKitProvider {...onchainKitProps}>
-          <div
-            style={{
-              paddingTop: context?.client.safeAreaInsets?.top ?? 0,
-              paddingBottom: context?.client.safeAreaInsets?.bottom ?? 0,
-              paddingLeft: context?.client.safeAreaInsets?.left ?? 0,
-              paddingRight: context?.client.safeAreaInsets?.right ?? 0,
-            }}
-          >
-            {children}
-          </div>
-        </OnchainKitProvider>
-      </DefaultOnchainKitProviders>
+      <AutoConnect enabled={autoConnect}>{children}</AutoConnect>
     </MiniKitContext.Provider>
+  );
+}
+
+export function MiniKitProvider({
+  children,
+  notificationProxyUrl,
+  enabled,
+  autoConnect,
+}: MiniKitProviderProps) {
+  if (!enabled) {
+    return children;
+  }
+
+  return (
+    <MiniKitProviderContent
+      notificationProxyUrl={notificationProxyUrl}
+      autoConnect={autoConnect}
+    >
+      {children}
+    </MiniKitProviderContent>
   );
 }