+++ packages/onchainkit/src/minikit/types.ts
@@ -1,27 +1,42 @@
-import type { Context, FrameNotificationDetails } from '@farcaster/frame-sdk';
+import type {
+  Context,
+  MiniAppNotificationDetails,
+} from '@farcaster/miniapp-sdk';
+import { PropsWithChildren } from 'react';
 
 export type UpdateClientContextParams = {
-  details?: FrameNotificationDetails | null;
-  frameAdded?: boolean;
+  details?: MiniAppNotificationDetails | null;
+  miniAppAdded?: boolean;
 };
 
-/**
- * Note: exported as public Type
- */
-export type MiniKitProviderReact = {
-  children: React.ReactNode;
-  /*
+export type MiniKitOptions = {
+  /**
+   * Whether to enable MiniKit features.
+   * @defaultValue `false`
+   */
+  enabled?: boolean;
+  /**
    * The URL of the notification proxy.
    * Notifications are sent by posting a cross origin request to a url returned by
    * the frames context.  This prop allows you to set a custom proxy route for MiniKits
    * notification related hooks to use.
-   * @default `/api/notify`
+   * @defaultValue `/api/notify`
    */
   notificationProxyUrl?: string;
+  /**
+   * Whether to automatically connect to the Farcaster connector if the user is in a Mini App.
+   *
+   * @defaultValue `true`
+   */
+  autoConnect?: boolean;
 };
 
+export type MiniKitProviderProps = PropsWithChildren<MiniKitOptions>;
+
 export type MiniKitContextType = {
-  context: Context.FrameContext | null;
+  enabled: boolean;
+  context: Context.MiniAppContext | null;
   updateClientContext: (params: UpdateClientContextParams) => void;
   notificationProxyUrl: string;
+  __isMiniKit: boolean;
 };